+++ src/Products/CMFPlomino/tests/test_doctest.py
@@ -23,6 +23,7 @@ def test_suite():
         'plomino_browser.txt',
         'plomino_file_attachment.txt',
         'plomino_view.txt',
+        'plomino_index.txt',
         # 'samples.txt',
         # 'plomino_usage.txt',
         # 'form-resources.txt',